pyproject.toml
[project]
name = "driving-world-model"
version = "0.1.0"
description = "Minimal R2-Dreamer reimplementation for DMC Vision"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.4",
    "torchrl>=0.9",
    "tensordict>=0.9",
    "hydra-core>=1.3",
    "dm-control>=1.0",
    "mujoco>=3.0",
    "gymnasium>=1.0",
    "numpy>=1.23",
    "opencv-python>=4.7",
    "tensorboard>=2.17",
    "einops>=0.3",
    "ruamel.yaml>=0.17",
    "moviepy>=1.0",
]
